requirements.txt
streamlit
google-generative
python-dotenv